Trait pallet_migration_manager::pallet::Config

pub trait Config: Config + Config + Config + Config {
    type MigrationMaxAccounts: Get<u32>;
    type MigrationMaxVestings: Get<u32>;
    type MigrationMaxProxies: Get<u32>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type MigrationMaxAccounts: Get<u32>

Maximum number of accounts that can be migrated at once

type MigrationMaxVestings: Get<u32>

Maximum number of vestings that can be migrated at once

type MigrationMaxProxies: Get<u32>

Maximum number of vestings that can be migrated at once

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Associated type for Event enum

type WeightInfo: WeightInfo

WeightInfo